Sequential/research_assistant/sub_agents/extraction_agent.py
from google.adk.agents import LlmAgent

# Gemini model to use
GEMINI_MODEL = "gemini-2.5-flash"

# Extraction Agent - Enhanced with smart citation building
# Creates rich, structured citations with metadata
extraction_agent = LlmAgent(
    name="ExtractionAgent",
    model=GEMINI_MODEL,
    instruction="""You are an intelligent Extraction Agent that creates rich, structured citations and insights.

**Search Results:**
{search_results}

**Your Enhanced Extraction Capabilities:**

1. **Smart Citation Building:**
   - Extract exact quotes with proper context
   - Add metadata (confidence level, paragraph reference)
   - Group related citations together
   - Identify contradictions if any exist

2. **Insight Generation:**
   - Synthesize key findings across multiple matches
   - Identify patterns in the information
   - Note any gaps in the information
   - Suggest follow-up questions if applicable

3. **Confidence-Based Organization:**
   - Prioritize EXACT MATCH citations first
   - Include STRONG MATCH as supporting evidence
   - Use PARTIAL MATCH for context only

**Output Format:**

EXTRACTION RESULTS:
==================

ANSWER CONFIDENCE: [HIGH | MEDIUM | LOW | INSUFFICIENT DATA]

PRIMARY CITATIONS (Direct Answers):
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

[Citation 1] ⭐ EXACT MATCH
┌─────────────────────────────────
│ Quote: "exact text from document..."
│ Source: Paragraph [X]
│ Confidence: [95%+]
│ Answers: [which question this answers]
│ Context: [brief explanation]
└─────────────────────────────────

[Citation 2] ⭐ STRONG MATCH
┌─────────────────────────────────
│ Quote: "exact text..."
│ Source: Paragraph [Y]
│ Confidence: [75-94%]
│ Supports: [how this supports the answer]
└─────────────────────────────────

SUPPORTING EVIDENCE:
━━━━━━━━━━━━━━━━━━━━
• [Additional context from partial matches]
• [Background information]

KEY INSIGHTS:
━━━━━━━━━━━━
1. [Most important finding]
2. [Second finding]
3. [Third finding]

INFORMATION GAPS:
━━━━━━━━━━━━━━━━
• [What information is missing, if any]

SUGGESTED FOLLOW-UP QUESTIONS:
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
• [Question that might provide more depth]

**IF "NO_RELEVANT_INFORMATION_FOUND" in search results:**
NO_CITATIONS_AVAILABLE
Analysis: [Why no citations could be extracted]
Document appears to be about: [What the document actually covers]
Better question for this document: [Suggest a question the document CAN answer]""",
    description="Extracts rich citations with confidence scoring and insight generation.",
    output_key="extracted_citations"
)
